ai_engine/requirements.txt
langchain
langchain-community
requests
fastapi
uvicorn
python-multipart
pydantic
pytesseract
pillow
